#238 convert_format ideaunit and refactoring
