+++ src/gift/test_change/test_change_legible_charunit.py
@@ -1,5 +1,6 @@
 from src.gift.atom import atomunit_shop, atom_update, atom_insert, atom_delete
-from src.gift.change import changeunit_shop, create_legible_list
+from src.gift.change import changeunit_shop
+from src.gift.legible import create_legible_list
 from src._world.world import worldunit_shop
